requirements.txt
streamlit
faiss-cpu
sentence-transformers
google-generativeai
python-dotenv
Pillow
numpy
